Split no_std/std-alloc UI tests
