+++ bon/tests/integration/ui/compile_fail/attr_into_future.stderr
@@ -1,83 +1,43 @@
 error: `#[builder(derive(IntoFuture(...)))` can only be used with async functions; using it with a synchronous function is likely a mistake
- --> tests/integration/ui/compile_fail/attr_into_future.rs:5:1
+ --> tests/integration/ui/compile_fail/attr_into_future.rs:4:1
   |
-5 | #[builder(derive(IntoFuture(Box)))]
+4 | #[builder(derive(IntoFuture(Box)))]
   | ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
   |
   = note: this error originates in the attribute macro `builder` (in Nightly builds, run with -Z macro-backtrace for more info)
 
 error: `#[builder(derive(IntoFuture(...)))` is not supported for unsafe functions because `IntoFuture::into_future()` method is a safe method
-  --> tests/integration/ui/compile_fail/attr_into_future.rs:12:7
+  --> tests/integration/ui/compile_fail/attr_into_future.rs:11:7
    |
-12 | async unsafe fn unsafe_async_function() -> u32 {
+11 | async unsafe fn unsafe_async_function() -> u32 {
    |       ^^^^^^
 
 error: `#[builder(derive(IntoFuture(...)))` is incompatible with `#[builder(finish_fn)]` members because `IntoFuture::into_future()` method accepts zero parameters
-  --> tests/integration/ui/compile_fail/attr_into_future.rs:18:35
+  --> tests/integration/ui/compile_fail/attr_into_future.rs:17:35
    |
-18 | async fn with_finish_fn(#[builder(finish_fn)] value: u32) -> u32 {
+17 | async fn with_finish_fn(#[builder(finish_fn)] value: u32) -> u32 {
    |                                   ^^^^^^^^^
 
 error: expected an attribute of form `IntoFuture(Box, ...)`
-  --> tests/integration/ui/compile_fail/attr_into_future.rs:23:18
+  --> tests/integration/ui/compile_fail/attr_into_future.rs:22:18
    |
-23 | #[builder(derive(IntoFuture))]
+22 | #[builder(derive(IntoFuture))]
    |                  ^^^^^^^^^^
 
 error: expected `Box` as the first argument, only boxed futures are supported
-  --> tests/integration/ui/compile_fail/attr_into_future.rs:29:29
+  --> tests/integration/ui/compile_fail/attr_into_future.rs:28:29
    |
-29 | #[builder(derive(IntoFuture(Arc)))]
+28 | #[builder(derive(IntoFuture(Arc)))]
    |                             ^^^
 
 error: expected `?Send` as the second argument
-  --> tests/integration/ui/compile_fail/attr_into_future.rs:35:34
+  --> tests/integration/ui/compile_fail/attr_into_future.rs:34:34
    |
-35 | #[builder(derive(IntoFuture(Box, Send)))]
+34 | #[builder(derive(IntoFuture(Box, Send)))]
    |                                  ^^^^
 
 error: `#[builder(derive(IntoFuture(...)))` can only be used with async functions; using it with a synchronous function is likely a mistake
-  --> tests/integration/ui/compile_fail/attr_into_future.rs:43:8
+  --> tests/integration/ui/compile_fail/attr_into_future.rs:42:8
    |
-43 | struct AsyncConfig {
+42 | struct AsyncConfig {
    |        ^^^^^^^^^^^
-
-error[E0277]: `dyn Future<Output = u32>` cannot be sent between threads safely
-  --> tests/integration/ui/compile_fail/attr_into_future.rs:60:17
-   |
-60 |     assert_send(&Sut.sut().value(21).into_future());
-   |                 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ `dyn Future<Output = u32>` cannot be sent between threads safely
-   |
-   = help: the trait `Send` is not implemented for `dyn Future<Output = u32>`
-   = note: required for `Unique<dyn Future<Output = u32>>` to implement `Send`
-note: required because it appears within the type `Box<dyn Future<Output = u32>>`
-  --> $RUST/alloc/src/boxed.rs
-   |
-   | pub struct Box<
-   |            ^^^
-note: required because it appears within the type `Pin<Box<dyn Future<Output = u32>>>`
-  --> $RUST/core/src/pin.rs
-   |
-   | pub struct Pin<Ptr> {
-   |            ^^^
-   = note: required for the cast from `&Pin<Box<dyn Future<Output = u32>>>` to `&dyn Send`
-
-error[E0277]: `(dyn Future<Output = u32> + 'static)` cannot be sent between threads safely
-  --> tests/integration/ui/compile_fail/attr_into_future.rs:67:17
-   |
-67 |     assert_send(&sut().value(21).into_future());
-   |                 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ `(dyn Future<Output = u32> + 'static)` cannot be sent between threads safely
-   |
-   = help: the trait `Send` is not implemented for `(dyn Future<Output = u32> + 'static)`
-   = note: required for `Unique<(dyn Future<Output = u32> + 'static)>` to implement `Send`
-note: required because it appears within the type `Box<(dyn Future<Output = u32> + 'static)>`
-  --> $RUST/alloc/src/boxed.rs
-   |
-   | pub struct Box<
-   |            ^^^
-note: required because it appears within the type `Pin<Box<(dyn Future<Output = u32> + 'static)>>`
-  --> $RUST/core/src/pin.rs
-   |
-   | pub struct Pin<Ptr> {
-   |            ^^^
-   = note: required for the cast from `&Pin<Box<(dyn Future<Output = u32> + 'static)>>` to `&dyn Send`
