+++ bon/tests/integration/ui/mod.rs
@@ -3,4 +3,9 @@
 fn ui() {
     let t = trybuild::TestCases::new();
     t.compile_fail("tests/integration/ui/compile_fail/*.rs");
+
+    if cfg!(any(feature = "std", feature = "alloc")) {
+        let t = trybuild::TestCases::new();
+        t.compile_fail("tests/integration/ui/compile_fail/std_or_alloc/*.rs");
+    }
 }
